Refine utest code.
